New method MQTTLib.ClientConnection.EasyPublish()
This method allows the code to send a message by providing a topic, a message, a QoS Level and optionally the retain flag in a single line.
It returns the packet ID used to send the packet. If the QoS level used is At Most Once Delivery, the packet ID will always be 0.